# api/openapi.yml
openapi: 3.1.0
info:
  title:  Aeneas API
  description: |
    This is the API documentation for  Aeneas, a go-lib
    created service.
  version: 0.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /example/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      summary: Example endpoint.
      description: |
        This is an example endpoint to demonstrate how to add custom API routes to the created service.
      responses:
        "200":
          description: Example response
          content:
            application/json:
              schema:
                type: object
                required:
                - id
                - name
                - description
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  description:
                    type: string